Fix the inability to discover multiple commissionable devices.
This issue happened when:

* There is more than one commissionable device online.
* Run `chip-tool discover commissionalbes --discover-once false`

We'll see if there will be only one result.
The root cause is that when the active resolve attempt receives the first result, it clears this attempt without waiting for more response.
The solution is recording when we get the response but keep waiting until it's timed out.